Revert "Game room: Players can join a specific game room"
