+++ app/components/GameList.jsx
@@ -1,11 +1,9 @@
 import React from 'react';
 import ReactDOM, { render } from 'react-dom';
-import { Link, NavLink, Router, withRouter } from 'react-router-dom';
+import { Link, NavLink, Router } from 'react-router-dom';
 import firebase from 'firebase';
 import { connect } from 'react-redux';
 import { showGameList } from '../reducers/conditionals';
-import { chooseGame } from '../reducers/auth';
-
 
 class GameList extends React.Component {
   constructor() {
@@ -13,7 +11,6 @@ class GameList extends React.Component {
     this.state = {
       games: false,
     };
-    this.selectGame = this.selectGame.bind(this)
   }
   componentWillMount() {
     const database = firebase.database();
@@ -26,16 +23,9 @@ class GameList extends React.Component {
     this.props.showGameList(false);
   }
 
-  selectGame(evt) {
-    let game = evt.target.name;
-    this.props.chooseGame(game);            
-    this.props.history.push(`/game/${game}/private`);    
-  }
-
   back = () => {
     this.props.showGameList(false);
   }
-
   render() {
     return (
       <div>
@@ -47,20 +37,18 @@ class GameList extends React.Component {
               <th><abbr title='Map'>Map</abbr></th>
               <th><abbr title='NumPlayers'>Player #</abbr></th>
               <th><abbr title='Security'>Private / Public</abbr></th>
-              <th><abbr title='Select'>Select</abbr></th>
             </tr>
           </thead>
           <tbody>
             {this.state.games &&
               Object.keys(this.state.games).map((game) => {
                 return (
-                  <tr key={game} onClick={(evt) => this.selectGame(evt)}>
+                  <tr key={game}>
                     <th>{game}</th>
                     <th>RoomName placeholder</th>
                     <th>map name placeholder</th>
                     <th>[{this.state.games[game].playersInGame.length}/4]</th>
                     <th>Public/Private</th>
-                    <th ><a name={game} className="button is-primary">JOIN</a></th>
                   </tr>
                 );
               })
@@ -69,13 +57,12 @@ class GameList extends React.Component {
         </table>
         <button className='button is-primary' id='neon' onClick={() => this.back()}>
           Back
-        </button>
+           </button>
       </div>
     );
   }
 }
 
-const mapDispatch = ({ showGameList, chooseGame });
+const mapDispatch = ({ showGameList });
 
-
-export default withRouter(connect(null, mapDispatch)(GameList));
+export default connect(null, mapDispatch)(GameList);
