+++ app/main.jsx
@@ -15,7 +15,7 @@ import FooterSection from './components/Footer';
 import Scores from './components/Scores';
 import ChooseBall from './components/ChooseBall';
 import Register from './components/Register';
-import GameWaitRoom from './components/GameWaitRoom';
+import PrivateGameRoom from './components/PrivateGameRoom';
 import ChooseGame from './components/ChooseGame';
 import GameType from './components/GameType';
 import GameList from './components/GameList';
@@ -50,7 +50,7 @@ const App = ({ children }) =>
           loginObj.uid
           ? <Game auth={auth} loginObj={loginObj}/>
           : <Redirect to ='/' />)}/>
-        <Route exact path="/game/:id/private" component={GameWaitRoom}/>
+        <Route exact path="/game/:id/private" component={PrivateGameRoom}/>
         <Route exact path="/choose" render={() => <GameType auth={auth} loginObj={loginObj}/>}/>
         <Route exact path="/game/:id/ball" render={() => <ChooseBall auth={auth} loginObj={loginObj}/>}/>
         <Route exact path="/scores" component={Scores}/>
