+++ public/styles.css
@@ -218,7 +218,3 @@ body {
 #choose {
   margin: 10% 10%;
 }
-
-#centerButtons {
-  padding-left: 37%;
-}
